refactor: extract admin bulk gear actions
